split handler functions to handler package.
