+++ types/types.go
@@ -1,4 +1,4 @@
-package main
+package types
 
 import "time"
